Shares some code between the two part of the problem.
